requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
geopy==2.3.0
pymongo==3.11.4
python_bcrypt==0.3.2
Requests==2.31.0
